user-sync/requirements.txt
python-gitlab
python-keycloak
requests
urllib3